Call stop on the parent queue service after we do our business.
JMSQueueService#stop marks the context as async, asynchronously does
what it needs to do, then marks the context as complete. If that
complete call happens before the shouldDestroy block in
DestroyableJMSQueueService#stop runs, the MSC can set the jsmServer
injector to null, causing that block to fail. This makes sure that
context isn't completed until the child does what it needs to do.